refactor: rename 'config' folder to 'store'
