+++ template-sample-app/template/src/components/views/Main.js
@@ -22,7 +22,7 @@ import Map from 'components/common/Map';
 import Legend from 'components/legends/Legend';
 import ZoomControl from 'components/common/ZoomControl';
 import { getLayers } from 'components/layers';
-import { setBottomSheetOpen, setError } from 'config/appSlice';
+import { setBottomSheetOpen, setError } from 'store/appSlice';
 import cartoLogoMap from 'assets/img/carto-logo-map.svg';
 
 const drawerWidth = 350;
