+++ template-sample-app/template/src/components/views/login/UserMenuLogin.js
@@ -3,7 +3,7 @@ import { Button, Hidden, IconButton } from '@material-ui/core';
 import AccountCircleOutlinedIcon from '@material-ui/icons/AccountCircleOutlined';
 import { useOAuthLogin } from '@carto/react/oauth';
 import { setTokenAndUserInfoAsync } from '@carto/react/redux';
-import { setError } from 'config/appSlice';
+import { setError } from 'store/appSlice';
 
 /**
  * Shows a login button.
