+++ template-sample-app/template/src/store/appSlice.js
