+++ template-sample-app/template/src/store/initialStateSlice.js
