+++ template-sample-app/template/src/store/store.js
