+++ template-skeleton/template/src/store/store.js
